Refactor terminal JournalingScreenCommands into reusable mixin class
